requirements.txt
requests==2.27.1
Adafruit-Blinka==7.1.1
adafruit-circuitpython-neopixel==6.2.4
adafruit-circuitpython-pixelbuf==1.1.2
Adafruit-PlatformDetect==3.22.0
Adafruit-PureIO==1.1.9
pyftdi==0.54.0
pyserial==3.5
pyusb==1.2.1
